Adjust favicons to force background colour
The actual favicon.ico icons and the SVG version can have transparency
because those are commonly used in browser tabbars and stuff.

The apple-touch-icon.png must not have transparency because that's not
supported on the iOS Home screen.

The other sizes are commonly used for PWA installations, which might be
used by iOS, so we'll ensure we have a solid background there too.

One of these days we'll add maskable PWA icons...